+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 Improvements:
 * [OLMIS-6332](https://openlmis.atlassian.net/browse/OLMIS-6332): Changed default page size to max integer.
 
+Bug fixes:
+* [OLMIS-6230](https://openlmis.atlassian.net/browse/OLMIS-6230): Fixed bug with null value of average consumption when column is not selected in template.
+
 8.0.0 / 2019-05-27
 ==================